removed FirebaseComponent and put it in main to fix replay Render
